Added an optional name event to the end of replay packets

+++ Challenges.h
@@ -32,7 +32,7 @@ struct Challenge {
 
 	std::list<Score> scores;
 
-	sf::Uint8 scoreCount;
+	sf::Uint16 scoreCount;
 
 	bool update;
